Get rid of `ecdsa.Node` type a push everything to `pkg/tbtc` package
