--- pkg/ecdsa/signer.go
@@ -1,7 +1,0 @@
-package ecdsa
-
-// ThresholdSigner is created from the threshold ECDSA DKG protocol and
-// represent a ready-to-work signing party.
-type ThresholdSigner struct {
-	// TODO: Implementation.
-}
