+++ pkg/tbtc/chain.go
@@ -1,25 +1,23 @@
 package tbtc
 
 import (
-	"github.com/keep-network/keep-core/pkg/ecdsa"
+	"github.com/keep-network/keep-core/pkg/sortition"
 	"github.com/keep-network/keep-core/pkg/subscription"
+	"math/big"
 )
 
 // Chain represents the interface that the TBTC module expects to interact
 // with the anchoring blockchain on.
 type Chain interface {
-	// OnNewWalletRegistered registers a callback that is invoked when an
-	// on-chain notification of the new wallet registration is seen.
-	OnNewWalletRegistered(
-		func(event *NewWalletRegisteredEvent),
-	) subscription.EventSubscription
+	// OnDKGStarted registers a callback that is invoked when an on-chain
+	// notification of the DKG process start is seen.
+	OnDKGStarted(func(event *DKGStartedEvent)) subscription.EventSubscription
 
-	ecdsa.Chain
+	sortition.Chain
 }
 
-// NewWalletRegisteredEvent represents an event informing about a new wallet
-// registration.
-type NewWalletRegisteredEvent struct {
-	WalletID         ecdsa.WalletID
-	WalletPubKeyHash [20]byte
+// DKGStartedEvent represents a DKG start event.
+type DKGStartedEvent struct {
+	Seed        *big.Int
+	BlockNumber uint64
 }
